+++ src/common/non_utf8/pathbuf.rs
@@ -686,3 +686,103 @@ where
         self.components().cmp(other.components())
     }
 }
+
+#[cfg(feature = "std")]
+impl<T> TryFrom<PathBuf<T>> for std::path::PathBuf
+where
+    T: for<'enc> Encoding<'enc>,
+{
+    type Error = PathBuf<T>;
+
+    /// Attempts to convert a [`PathBuf`] into a [`std::path::PathBuf`], returning a result
+    /// containing the new path when successful or the original path when failed
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use std::convert::TryFrom;
+    /// use std::path::PathBuf;
+    /// use typed_path::UnixPathBuf;
+    ///
+    /// let unix_path_buf = UnixPathBuf::from("/path/to/file.txt");
+    /// let std_path_buf: PathBuf = TryFrom::try_from(unix_path_buf).unwrap();
+    /// ```
+    fn try_from(path: PathBuf<T>) -> Result<Self, Self::Error> {
+        match std::str::from_utf8(path.as_bytes()) {
+            Ok(s) => Ok(std::path::PathBuf::from(s)),
+            Err(_) => Err(path),
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+impl<T> TryFrom<std::path::PathBuf> for PathBuf<T>
+where
+    T: for<'enc> Encoding<'enc>,
+{
+    type Error = std::path::PathBuf;
+
+    /// Attempts to convert a [`std::path::PathBuf`] into a [`PathBuf`], returning a result
+    /// containing the new path when successful or the original path when failed
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use std::convert::TryFrom;
+    /// use std::path::PathBuf;
+    /// use typed_path::UnixPathBuf;
+    ///
+    /// let std_path_buf = PathBuf::from("/path/to/file.txt");
+    /// let unix_path_buf: UnixPathBuf = TryFrom::try_from(std_path_buf).unwrap();
+    /// ```
+    fn try_from(path: std::path::PathBuf) -> Result<Self, Self::Error> {
+        match path.to_str() {
+            Some(s) => Ok(PathBuf::from(s)),
+            None => Err(path),
+        }
+    }
+}
+
+#[cfg(any(
+    unix,
+    all(target_vendor = "fortanix", target_env = "sgx"),
+    target_os = "solid_asp3",
+    target_os = "hermit",
+    target_os = "wasi"
+))]
+#[cfg(feature = "std")]
+mod std_conversions {
+    use std::ffi::{OsStr, OsString};
+    #[cfg(all(target_vendor = "fortanix", target_env = "sgx"))]
+    use std::os::fortanix_sgx as os;
+    #[cfg(target_os = "solid_asp3")]
+    use std::os::solid as os;
+    #[cfg(any(target_os = "hermit", unix))]
+    use std::os::unix as os;
+    #[cfg(target_os = "wasi")]
+    use std::os::wasi as os;
+
+    use os::ffi::{OsStrExt, OsStringExt};
+
+    use super::*;
+
+    impl<T> From<PathBuf<T>> for OsString
+    where
+        T: for<'enc> Encoding<'enc>,
+    {
+        #[inline]
+        fn from(path_buf: PathBuf<T>) -> Self {
+            OsStringExt::from_vec(path_buf.into_vec())
+        }
+    }
+
+    impl<T> AsRef<OsStr> for PathBuf<T>
+    where
+        T: for<'enc> Encoding<'enc>,
+    {
+        #[inline]
+        fn as_ref(&self) -> &OsStr {
+            OsStrExt::from_bytes(self.as_bytes())
+        }
+    }
+}
